+++ README.md
@@ -1,5 +1,16 @@
-@GetMapping("/reply/{id}") 구현하기
-첨부파일, 
-댓글 달기, 
-좋아요/싫어요, 
-웹 에디터 달기
+1.게시판
+- 게시판 등록,조회,수정,삭제
+- 페이징  
+- 검색 
+- 댓글 등록,조회,수정,삭제
+2. 로그인,회원가입
+- 회원가입
+  - sns로그인
+  - 이메일 인증
+- 로그인
+- 비밀번호 찾기
+2. 사이드 프로젝트 
+- 등록,조회,수정,삭제
+- 검색
+3. 멤버 & 스터디
+- 등록,조회,수정,삭제
