+++ src/main/java/board/controller/board/BoardController.java
@@ -1,20 +1,18 @@
-package board.controller;
+package board.controller.board;
 
-import board.domain.Reply;
-import board.dto.BoardDto;
-import board.service.BoardService;
+import board.domain.board.Reply;
+import board.dto.board.BoardDto;
+import board.service.board.BoardService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
 import org.springframework.data.web.PageableDefault;
-import org.springframework.http.MediaType;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
 
 import java.util.List;
