+++ src/main/java/board/dto/board/BoardDto.java
@@ -1,11 +1,11 @@
-package board.dto;
+package board.dto.board;
 
-import board.domain.Board;
-import board.domain.Reply;
-import lombok.*;
+import board.domain.board.Board;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
 
 import java.time.LocalDateTime;
-import java.util.List;
 
 @Data
 @NoArgsConstructor
