+++ src/main/java/board/repository/board/UserRepository.java
@@ -1,6 +1,6 @@
-package board.repository;
+package board.repository.board;
 
-import board.domain.User;
+import board.domain.board.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface UserRepository extends JpaRepository<User, Long> {
